refactor(host-adapters): true adapter pattern, drop plugin-sdk re-exports
Consumers (extension, mobile, CLI) now import protocol types and HostCore
directly from @tlsn/plugin-sdk, and only platform glue from @tlsn/host-*.
This mirrors react + react-dom: the adapter lives alongside the SDK rather
than wrapping it.

- @tlsn/host-contracts: stop re-exporting Handler / PluginConfig /
  InterceptedRequest(Header) / ProveProgressData / RevealRangeDescriptor.
- @tlsn/host-react-native: stop re-exporting DomJson / Handler /
  PluginConfig / WindowMessage / InterceptedRequest(Header) /
  translateHandler(s) / NativeHandler / ApprovalMode. ApprovalMode now
  comes from @tlsn/host-contracts; the local duplicate is removed.
- @tlsn/host-cli: re-route eight files' protocol-type imports from
  host-contracts to plugin-sdk (event-emitter, resolve-plugin, eight
  adapter sources + index.test).
- app/mobile: add @tlsn/plugin-sdk as a direct dependency; split
  PluginScreen's import into host-rn (MobilePluginHost/EventEmitter/
  RevealRangeDescriptor) + plugin-sdk (PluginConfig/WindowMessage) +
  host-contracts (NativeHandler). registry.ts pulls PluginConfig from
  plugin-sdk.
