+++ app/mobile/package.json
@@ -25,6 +25,7 @@
     "@react-navigation/native": "^7.1.8",
     "@tlsn/host-contracts": "*",
     "@tlsn/host-react-native": "*",
+    "@tlsn/plugin-sdk": "*",
     "@tlsn/plugins": "*",
     "@types/uuid": "^10.0.0",
     "expo": "~54.0.33",
